linux week2 done

+++ c/rand/random.c
@@ -7,13 +7,13 @@
    Only effective if N is much smaller than RAND_MAX;
    if this may not be the case, use a better random
    number generator. */
-void shuffle(int *array, size_t n) {
+void shuffle(int *array, int n) {
     if (n > 1)
     {
-        size_t i;
+        int i;
         for (i = 0; i < n - 1; i++)
         {
-          size_t j = i + rand() / (RAND_MAX / (n - i) + 1);
+          int j = i + rand() / (RAND_MAX / (n - i) + 1);
           int t = array[j];
           array[j] = array[i];
           array[i] = t;
