+++ c/rand/random.h
@@ -1,5 +1,5 @@
 #ifndef RANDOM_H_
 #define RANDOM_H_
 
-void shuffle(int *array, size_t n);
+void shuffle(int *array, int n);
 #endif /* RANDOM_H_ */
